Adds additional function benchmarks
