+++ test/benchmark/data/SumMultiplesThreeFiveNaive.ek
@@ -1,0 +1,23 @@
+//import <test>
+
+func computeSumOfMultiplesOfThreeOrFiveBelowOneThousand()
+{
+    int sum = 0;
+
+    for (int i = 1; i < 1000; ++i)
+    {
+        if (i % 3 == 0)
+        {
+            sum = sum + i;
+        }
+        else if (i % 5 == 0)
+        {
+            sum = sum + i;
+        }
+    }
+
+    return sum;
+}
+
+
+//TEST(computeSumOfMultiplesOfThreeOrFiveBelowOneThousand() == 233168, "");
