+++ test/benchmark/data/SumMultiplesThreeFiveOpt.ek
@@ -1,0 +1,26 @@
+//import <test>
+
+func computeSumOfMultiplesOfThreeOrFiveBelowOneThousand()
+{
+    int sum = 0;
+
+    for (int i = 15; i < 1000; i = i + 15)
+    {
+        sum = (sum - i);
+    }
+
+    for (int i = 3; i < 1000; i = i + 3)
+    {
+        sum = (sum + i);
+    }
+
+    for (int i = 5; i < 1000; i = i + 5)
+    {
+        sum = (sum + i);
+    }
+
+    return sum;
+}
+
+
+//TEST(computeSumOfMultiplesOfThreeOrFiveBelowOneThousand() == 233168, "");
